// ui/src/api/apiSchemas.ts
/**
 * Generated by @openapi-codegen
 *
 * @version 1.0
 */
export type ChangePasswordIN = {
  currentPassword: string;
  newPassword: string;
};

export type ChangePasswordOUT = {
  apiKey: string;
};

export type ErrorOUT = {
  error: string;
};

export type ForgotPasswordIN = {
  loginOrEmail: string;
};

export type ForgotPasswordOUT = Record<string, any>;

export type GetUserOUT = {
  login: string;
  email: string;
  /**
   * @format date-time
   */
  createdOn: string;
};

export type LoginIN = {
  loginOrEmail: string;
  password: string;
  /**
   * @format int32
   */
  apiKeyValidHours?: number;
};

export type LoginOUT = {
  apiKey: string;
};

export type LogoutIN = {
  apiKey: string;
};

export type LogoutOUT = Record<string, any>;

export type PasswordResetIN = {
  code: string;
  password: string;
};

export type PasswordResetOUT = Record<string, any>;

export type RegisterIN = {
  login: string;
  email: string;
  password: string;
};

export type RegisterOUT = {
  apiKey: string;
};

export type UpdateUserIN = {
  login: string;
  email: string;
};

export type UpdateUserOUT = Record<string, any>;

export type VersionOUT = {
  buildSha: string;
};